Migrated to .NET Core
Also updated nUnit to 3.0 and added test adapter NuGet package to run tests within Visual Studio
