+++ FastString.Test/Utf8WriterTest.cs
@@ -21,41 +21,41 @@ namespace FastString.Test
 		[Test]
 		public void AppendFormatSimple()
 		{
-			writer.AppendFormat(new utf8("Hello {0}!"), "world");
-			Assert.That(Bake(), Is.EqualTo(new utf8("Hello world!")));
+			writer.AppendFormat(new Utf8String("Hello {0}!"), "world");
+			Assert.That(Bake(), Is.EqualTo(new Utf8String("Hello world!")));
 		}
 
 		[Test]
 		public void AppendFormatNumber()
 		{
-			writer.AppendFormat(new utf8("Hello {0}!"), 12);
-			Assert.That(Bake(), Is.EqualTo(new utf8("Hello 12!")));
+			writer.AppendFormat(new Utf8String("Hello {0}!"), 12);
+			Assert.That(Bake(), Is.EqualTo(new Utf8String("Hello 12!")));
 		}
 
 		[Test]
 		public void AppendFormatNumberWithOptions()
 		{
-			writer.AppendFormat(new utf8("Hello {0:X4}!"), 12);
-			Assert.That(Bake(), Is.EqualTo(new utf8("Hello 000C!")));
+			writer.AppendFormat(new Utf8String("Hello {0:X4}!"), 12);
+			Assert.That(Bake(), Is.EqualTo(new Utf8String("Hello 000C!")));
 		}
 
 		[Test]
 		public void AppendFormatSeveralArgs()
 		{
-			writer.AppendFormat(new utf8("Hello {1}, {0:X4} times {2}!"), 12, "world", new MockToString("over"));
-			Assert.That(Bake(), Is.EqualTo(new utf8("Hello world, 000C times over!")));
+			writer.AppendFormat(new Utf8String("Hello {1}, {0:X4} times {2}!"), 12, "world", new MockToString("over"));
+			Assert.That(Bake(), Is.EqualTo(new Utf8String("Hello world, 000C times over!")));
 		}
 
 		[Test]
 		public void AppendFormatEscape()
 		{
-			writer.AppendFormat(new utf8("Hello {{0}{0}!"), "world");
-			Assert.That(Bake(), Is.EqualTo(new utf8("Hello {0}world!")));
+			writer.AppendFormat(new Utf8String("Hello {{0}{0}!"), "world");
+			Assert.That(Bake(), Is.EqualTo(new Utf8String("Hello {0}world!")));
 		}
 
-		utf8 Bake()
+		Utf8String Bake()
 		{
-			var b = new utf8(
+			var b = new Utf8String(
 				new ArraySegment<byte>(stream.GetBuffer(), 0, (int)stream.Position));
 			Console.WriteLine(b.Length);
 			Console.WriteLine("[{0}]", b);
