+++ FastString.Test/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="NUnit" version="2.6.4" targetFramework="net45" />
+  <package id="NUnit" version="3.6.1" targetFramework="net451" />
+  <package id="NUnit3TestAdapter" version="3.7.0" targetFramework="net451" />
 </packages>
