+++ FastString.sln
@@ -1,6 +1,7 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
+# Visual Studio 15
+VisualStudioVersion = 15.0.26403.3
+MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FastString", "FastString\FastString.csproj", "{08E0F5AB-063A-47ED-9E72-211E0C999098}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FastString.Test", "FastString.Test\FastString.Test.csproj", "{AEF8548A-45B0-4F9B-A099-D4C88049282E}"
@@ -25,6 +26,9 @@ Global
 		{98665FEC-78C1-4F57-B84F-D3F6972F02DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{98665FEC-78C1-4F57-B84F-D3F6972F02DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{98665FEC-78C1-4F57-B84F-D3F6972F02DE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		Policies = $0
