+++ FastString/FastString.csproj
@@ -1,52 +1,25 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{08E0F5AB-063A-47ED-9E72-211E0C999098}</ProjectGuid>
-    <OutputType>Library</OutputType>
-    <RootNamespace>FastString</RootNamespace>
-    <AssemblyName>FastString</AssemblyName>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
+    <TargetFrameworks>netstandard1.3;net451</TargetFrameworks>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-    <DebugType>full</DebugType>
-    <Optimize>false</Optimize>
-    <OutputPath>bin\Debug</OutputPath>
-    <DefineConstants>DEBUG;</DefineConstants>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <Optimize>true</Optimize>
-    <OutputPath>bin\Release</OutputPath>
-    <ErrorReport>prompt</ErrorReport>
-    <WarningLevel>4</WarningLevel>
-    <ConsolePause>false</ConsolePause>
-  </PropertyGroup>
+
   <ItemGroup>
-    <Reference Include="System" />
+    <None Remove="Unicode\chardata" />
+    <None Remove="Unicode\charnames" />
   </ItemGroup>
-  <ItemGroup>
-    <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="StringSlice.cs" />
-    <Compile Include="utf8.cs" />
-    <Compile Include="UtfIndex.cs" />
-    <Compile Include="ReverseUtf8Enumerator.cs" />
-    <Compile Include="Utf8Enumerator.cs" />
-    <Compile Include="Utf8Builder.cs" />
-    <Compile Include="Utf8Writer.cs" />
-    <Compile Include="Splitter.cs" />
-    <Compile Include="Unicode\CharInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Unicode\" />
-  </ItemGroup>
+
   <ItemGroup>
     <EmbeddedResource Include="Unicode\chardata" />
     <EmbeddedResource Include="Unicode\charnames" />
   </ItemGroup>
-  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
+
+  <ItemGroup>
+    <PackageReference Include="System.Diagnostics.Contracts" Version="4.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Properties\" />
+  </ItemGroup>
+
 </Project>
